+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes
 ## UNRELEASED
 
 * Bug: Handle special characters in filenames like `[...id].js` during tracing.
+* Deps: Upgrade `globby` to `v11`.
 
 ## 0.10.8