+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changes
 =======
 
-## UNRELEASED
+## 0.10.9
 
 * Bug: Handle special characters in filenames like `[...id].js` during tracing.
 * Deps: Upgrade `globby` to `v11`.